norme stuff here and there

+++ src/vm/stdout_messages.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   stdout_messages.c                                  :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: jaelee <jaelee@student.42.fr>              +#+  +:+       +#+        */
+/*   By: aamadori <aamadori@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/05/07 18:39:17 by jaelee            #+#    #+#             */
-/*   Updated: 2019/05/07 19:05:14 by jaelee           ###   ########.fr       */
+/*   Updated: 2019/05/21 15:02:42 by aamadori         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,7 +22,7 @@ void	print_prelude(t_vm_state *state)
 	players = state->players;
 	i = 0;
 	ft_printf("Introducing the contestants...\n");
-	while (i < state->players.length)
+	while (i < (int)state->players.length)
 	{
 		ft_printf("PLAYER %d, weighing %u bytes, ",
 			i + 1, ((t_player*)players.ptr)[i].header.prog_size);
